third_party/WebKit/LayoutTests/fast/layout/border-image-outset-no-relayout.html

<!DOCTYPE html>
<style>
    div {
        border-width: 21px 30px 30px 21px;
        width: 75px;
        height: 75px;
        display: inline-block;
        border-image-source: url("../borders/resources/border-image.png");
        border-image-slice: 21 30 30 21 fill;
        border-image-width: 1;
        border-image-outset: 1;
    }
</style>
<script src="../../resources/js-test.js"></script>
<div id="test"></div>
<script>
// Force style recalc and layout.
document.body.offsetTop;

if (window.internals)
    shouldBe("internals.needsLayoutCount()", "0", true);

// Changing border-image-outset will cause a repaint but no layout.
document.getElementById("test").style['border-image-outset'] = '5';

if (window.internals) {
    shouldBe("internals.needsLayoutCount()", "0", true);
}
</script>
